requirements.txt
streamlit
PyPDF2
fpdf
python-dotenv
google-generativeai
